Eddy 2.1.1
Cleaned up a lot of code and fixed a few potential bugs

Source code security is at an all-time high. This release seperates the core, your app code and public resources; making for easier updates of the core and safer business logic

The core can now be centralised. So you don't need a copy of the core for each app; just one central version (you just need to update the /public/index.php file to point to the right location of the /core/init.php

Better handling of models:
- Store original data in the object for reference/rollback
- Better handling of different datatypes
- Updates only send the data that's changed

Better first-run content