Fixed a bug with the new routing mechanism
This is a crafty one and may crop up again… it affects requests with extensions (e.g. .php or .json)

Extensions allow you to choose a Response handler method. The default Response class handles JSON and XML in a very basic way.

However Routes were mangling the extension. This commit attempts to resolve that problem in most cases.